airbyte-integrations/connectors/source-apache-kafka/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-apache-kafka"
description = "Source implementation for apache-kafka."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/apache-kafka"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_apache_kafka" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
airbyte-cdk = "^6.33.0"
confluent-kafka = "2.4.0"
mockafka-py = "^0.1.63"
pytest = "^8.3.5"
freezegun = "^1.5.1"


[tool.poetry.scripts]
source-apache-kafka = "source_apache_kafka.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"

